refactor(engine): move inclusion list handling into its own module
`tree/mod.rs` had grown by ~480 lines of EIP-7805 code. The appendability
check, the retained-list store and their tests now live in
`tree/inclusion_list.rs`, leaving mod.rs a net 17 lines smaller than baseline.
